+++ apps/api/Program.cs
@@ -19,6 +19,7 @@ using Mohandseto.Api.Application.Engagement;
 using Mohandseto.Api.Application.Admin;
 using Mohandseto.Api.Application.AdminOrders;
 using Mohandseto.Api.Application.AdminQuotes;
+using Mohandseto.Api.Application.AdminContent;
 using Mohandseto.Api.Infrastructure;
 using Serilog;
 
@@ -87,6 +88,8 @@ builder.Services.AddScoped<SettingsService>();
 builder.Services.AddScoped<AdminDashboardService>();
 builder.Services.AddScoped<AdminOrderService>();
 builder.Services.AddScoped<AdminQuoteService>();
+builder.Services.AddScoped<AdminContentService>();
+builder.Services.AddHostedService<ContentDispatchWorker>();
 builder.Services.AddScoped<CustomizationService>();
 builder.Services.AddSingleton<ISmsSender, ConsoleSmsSender>();